Fase 2/Evidencias proyecto/Classflex/requirements.txt
asgiref==3.8.1
Django==5.0.8
django-cors-headers==4.4.0
djangorestframework==3.15.2
mssql-django==1.5
PyJWT==2.9.0
pyodbc==5.1.0
pytz==2024.1
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1